+++ tokenmagic/fx/glsl/fragmentshaders/dropshadow.js
@@ -1,4 +1,5 @@
 export const dropShadow = `
+precision mediump float;
 varying vec2 vTextureCoord;
 uniform sampler2D uSampler;
 uniform float alpha;